# [i18n] - Do not watch translations from gems when reloading is enabled